Audio first iteration
How audio works currently.
1. drag and drop the AudioManager prefab into a scene. And attach the
physical and spiritual cameras to it. The manager uses whichever camera
is enabled to tell if it should spawn a spirit or physical echo.

2. Add the AudioSpawn script to any object that should spawn sounds.

3. Fill in the clip and transform variables in the AudioSpawn script

4. Make a script or edit a script to call the spawn function in
AudioSpawn. For example, if you wanted to spawn a stalagtite dripping
you would need to make a stalagtiteDrip script that will call spawn
every 12.5 seconds. You can pass in a variable to say how long the sound
should stay, otherwise it stays for 10 seconds.